refactor : Input & Output 분리
